+++ bin/generate-tests.php
@@ -81,7 +81,6 @@ add_test(['context:context', '--context', 'run'], 'ContextContextRunTest');
 add_test(['context:context', '--context', 'my_default', '-vvv'], 'ContextContextMyDefaultTest');
 add_test(['context:context', '--context', 'no_no_exist'], 'ContextContextDoNotExistTest');
 add_test(['context:context', '--context', 'production'], 'ContextContextProductionTest');
-add_test(['parallel:sleep', '--sleep5', '0', '--sleep7', '0', '--sleep10', '0'], 'ParallelSleepTest');
 add_test([], 'NewProjectTest', '/tmp');
 add_test(['unknown:command', 'toto', '--foo', 1], 'NoConfigTest', '/tmp');