+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1 (not released yet)
 
+* Better handling of option without value
 * Fixed the stubs generation when castor is installed via composer
 * Fixed the initial castor.php file generated for new projects